requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
dataclasses-json==0.6.4
exceptiongroup==1.2.0
faiss-cpu==1.8.0
frozenlist==1.4.1
greenlet==3.0.3
idna==3.6
jsonpatch==1.33
jsonpointer==2.4
langchain-community==0.0.29
langchain-core==0.1.33
langchain-nvidia-ai-endpoints==0.0.4
langchain-text-splitters==0.0.1
langsmith==0.1.31
marshmallow==3.21.1
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
orjson==3.9.15
packaging==23.2
pillow==10.2.0
pydantic==2.6.4
pydantic_core==2.16.3
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.31.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.28
tenacity==8.2.3
typing-inspect==0.9.0
typing_extensions==4.10.0
urllib3==2.2.1
yarl==1.9.4
